Science Arm Code Integration

+++ jetson/ra_kinematics/arm_state.hpp
@@ -39,6 +39,7 @@ private:
 
             rot_axis << joint_geom["axis"][0], joint_geom["axis"][1], joint_geom["axis"][2];
 
+            // TODO check with RA whether joint limits are correct
             joint_limits.push_back(joint_geom["limit"]["lower"]);
             joint_limits.push_back(joint_geom["limit"]["upper"]);
 
@@ -190,7 +191,7 @@ public:
 
     bool obstacle_free();
 
-    int num_joints() const;
+    size_t num_joints() const;
 
     std::string get_child_link(size_t joint_index) const;
